Feat: Navigation to Spot Detail complete
